Fix linking errors and move headers outside of src folder - Fixed linker errors with action factory & utils sleep function - Moved include folder outside of src and updated cmake
