+++ engine-computer/include/mach/config_mapping.hpp
@@ -1,5 +1,4 @@
 #pragma once
-
 #include <vector>
 #include <string>
