+++ engine-computer/include/mach/sequences/action.hpp
@@ -1,5 +1,4 @@
 #pragma once
-
 #include <yaml-cpp/yaml.h>
 
 namespace mach {
