+++ engine-computer/include/mach/sequences/close_action.hpp
@@ -1,14 +1,13 @@
 #pragma once
-
 #include <memory>
-#include "./action.hpp"
+#include "action.hpp"
 #include "../device/valve.hpp"
 
 namespace mach {
     
-class OpenAction : public Action {
+class CloseAction : public Action {
     public:
-        OpenAction() : Action() {}
+        CloseAction() : Action() {}
 
         bool init(YAML::Node node);
         void execute() override;
