+++ engine-computer/include/mach/sequences/sleep_action.hpp
@@ -1,20 +1,19 @@
 #pragma once
-
 #include <memory>
-#include "./action.hpp"
+#include "action.hpp"
 #include "../device/valve.hpp"
 
 namespace mach {
     
-class CloseAction : public Action {
+class SleepAction : public Action {
     public:
-        CloseAction() : Action() {}
+        SleepAction() : Action() {}
 
         bool init(YAML::Node node);
         void execute() override;
 
     private:
-        std::shared_ptr<Valve> valve;
+        double duration;
 };
 
 }
