+++ engine-computer/include/mach/util.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include <vector>
 #include <string>
+#include <chrono>
 
 namespace mach::util {
 
@@ -8,7 +9,6 @@ const double* vectorToDouble(const std::vector<double> &doubleVector);
 
 const char** vectorToChar(const std::vector<std::string> &stringVector);
 
-template<typename Rep, typename Period>
-bool sleepOrAbort(std::chrono::duration<Rep, Period> duration);
+bool sleepOrAbort(std::chrono::milliseconds duration);
 
 } // namespace mach::util
