+++ engine-computer/src/util.cpp
@@ -1,6 +1,5 @@
-#include <chrono>
+#include "mach/util.hpp"
 #include <thread>
-#include "include/util.hpp"
 
 const double* mach::util::vectorToDouble(const std::vector<double> &doubleVector) {
     return doubleVector.data();
@@ -14,12 +13,11 @@ const char** mach::util::vectorToChar(const std::vector<std::string> &stringVect
     return charPtrArray;
 }
 
-template<typename Rep, typename Period>
-bool mach::util::sleepOrAbort(std::chrono::duration<Rep, Period> duration) {
+bool mach::util::sleepOrAbort(std::chrono::milliseconds duration) {
     // TODO Need software abort?
     // std::unique_lock<std::mutex> lock(abort_mutex);
     // abort_cv.wait_for(lock, duration, []() { return abort_flag.load(); });
     // return abort_flag.load();
     std::this_thread::sleep_for(duration);
-    return false;
+    return true;
 }
